requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.0
Flask-Migrate==4.0.0
psycopg2-binary==2.9.6
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1